+++ projects/ngx-scrollbar/src/lib/ng-scrollbar.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { BidiModule } from '@angular/cdk/bidi';
 import { PortalModule } from '@angular/cdk/portal';
 import { PlatformModule } from '@angular/cdk/platform';
-import { SmoothScrollModule } from 'ngx-scrollbar/smooth-scroll';
+import { SmoothScrollModule } from '@metrica-sports/ngx-scrollbar/smooth-scroll';
 
 import { NgScrollbar } from './ng-scrollbar';
 import { HideNativeScrollbar } from './hide-native-scrollbar/hide-native-scrollbar';